fix: explicit int8_t cast for extraLives comparison
Cast extraLives to int8_t before >= 0 check to ensure correct
signed comparison regardless of platform char signedness.

+++ src/GameManager.cpp
@@ -269,7 +269,7 @@ ChainCallbackResult GameManager::OnUpdate(GameManager *gameManager)
             gameManager->nextScoreIncrement = 0;
             gameManager->guiScore = gameManager->score;
         }
-        if (gameManager->extraLives >= 0 && g_ExtraLivesScores[gameManager->extraLives] <= gameManager->guiScore)
+        if ((int8_t)gameManager->extraLives >= 0 && g_ExtraLivesScores[gameManager->extraLives] <= gameManager->guiScore)
         {
             const bool hasSecondPlayer = HasSecondPlayer();
             if (gameManager->livesRemaining < MAX_LIVES || (hasSecondPlayer && gameManager->livesRemaining2 < MAX_LIVES))
